+++ scripts/core/helpers/uuid.js
